requirements.txt
python-telegram-bot == 12.7
APScheduler == 3.6.3